% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.root.r
\name{square.root}
\alias{square.root}
\title{Computes a square root}
\usage{
square.root(x)
}
\arguments{
\item{x}{a double vector of length 1}
}
\value{
The square root of `x`
}
\description{
Computes a square root
}
\details{
This function is for pedagogical illustration only. Please use
`base::sqrt` in R or `sqrt` in C++.
}
\examples{
square.root(2)

}
